Move setup logic to Bootstrap directory
Relocated the setup function and related code from Sources/Scout/Core/Setup/Setup.swift to Sources/Scout/Core/Bootstrap/Setup.swift for improved project structure and clarity. No functional changes were made.
