--- Sources/Scout/Core/Setup/Setup.swift
@@ -1,21 +1,0 @@
-//
-// Copyright 2025 Mikhail Kasianov
-//
-// Use of this source code is governed by an MIT-style
-// license that can be found in the LICENSE file or at
-// https://opensource.org/licenses/MIT.
-
-import CloudKit
-import Logging
-import Metrics
-
-@MainActor
-var container: CKContainer?
-
-@MainActor
-public func setup(container: CKContainer) throws {
-    Scout.container = container
-    LoggingSystem.bootstrap(CKLogHandler.init)
-    MetricsSystem.bootstrap(TelemetryFactory())
-    try NotificationListener.activity.setup()
-}
